move test objects to dedicated folder
