+++ packages/example-v8/src/stories/WithTypedProps.spec.ts
@@ -1,6 +1,6 @@
 import {test} from "@playwright/test";
+import {AppObject} from "../tests/objects/AppObject";
 import {assertStorybookIsRunning} from "../tests/utils";
-import {AppObject} from "../tests/utils/AppObject";
 import {TEST_TIMEOUT_MS} from "../tests/utils/constants";
 
 test.beforeAll(assertStorybookIsRunning);
