+++ packages/example-v8/src/tests/objects/ArgsTableObject.ts
